Simplified endpoints setup
